Refactor OverriddenAssignmentCleanUp to jdt.core.manipulation (#1371)
- create new OverriddenAssignmentCleanUpCore and
  OverriddenAssignmentFixCore classes in jdt.core.manipulation
- modify OverriddenAssignmentCleanUp to use core class
- remove old OverriddenAssignmentCleanUp
- fixes #1370
